+++ task/executor/k8s/client/informers/externalversions/generic.go
@@ -8,7 +8,7 @@ package externalversions
 import (
 	fmt "fmt"
 
-	v1 "github.com/bborbe/agent/task/executor/k8s/apis/agents.bborbe.dev/v1"
+	v1 "github.com/bborbe/agent/task/executor/k8s/apis/agent.benjamin-borbe.de/v1"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
 	cache "k8s.io/client-go/tools/cache"
 )
@@ -39,9 +39,9 @@ func (f *genericInformer) Lister() cache.GenericLister {
 // TODO extend this to unknown resources with a client pool
 func (f *sharedInformerFactory) ForResource(resource schema.GroupVersionResource) (GenericInformer, error) {
 	switch resource {
-	// Group=agents.bborbe.dev, Version=v1
-	case v1.SchemeGroupVersion.WithResource("agentconfigs"):
-		return &genericInformer{resource: resource.GroupResource(), informer: f.Agents().V1().AgentConfigs().Informer()}, nil
+	// Group=agent.benjamin-borbe.de, Version=v1
+	case v1.SchemeGroupVersion.WithResource("configs"):
+		return &genericInformer{resource: resource.GroupResource(), informer: f.Agent().V1().Configs().Informer()}, nil
 
 	}